rearrange: create rn module
